% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CLL_covariates}
\alias{CLL_covariates}
\title{CLL_covariates}
\format{A  data frame diagnosis and gender for the n=200 patient samples in CLL_data}
\usage{
data(CLL_covariates)
}
\description{
Data frame containing additional information on the patient samples, i.e. diagnosis and gender.
}
\keyword{datasets}
