% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CLL_data}
\alias{CLL_data}
\title{CLL_data}
\format{A list of matrices, with the following -omics:
\itemize{
  \item{mRNA: }{normalized expression values for the 5000 most variable genes}
  \item{Methylation: }{methylation M-values for the 4248 most variable CpG sites}
  \item{Drugs: }{viability values in response to 310 different drugs and concentrations}
  \item{Mutations: }{Mutation status for 69 selected genes}
}}
\source{
\url{\code{https://www.ncbi.nlm.nih.gov/pubmed/29227286}}
}
\usage{
data(CLL_data)
}
\description{
A list containing different omic measurements for Chronic lymphocytic leukaemia (CLL) patient samples. \cr
The data set was originally published in \url{https://www.ncbi.nlm.nih.gov/pubmed/29227286}.\cr
The MOFA analysis for this data set can be found in \url{http://msb.embopress.org/content/14/6/e8124}
}
\keyword{datasets}
