% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSigDB_v6.0_C5_human}
\alias{MSigDB_v6.0_C5_human}
\title{MSigDB_v6.0_C5_human}
\format{Rows contain biological pathways and columns contain genes}
\usage{
data(MSigDB_v6.0_C5_human)
}
\description{
A matrix containing human gene sets from the MSigDB 6.0 data base (http://software.broadinstitute.org/gsea/msigdb).
\describe{
\item{specie:}{Homo sapiens}
\item{Gene IDs:}{Ensembl IDs from version 75 (i.e. ENSG00000125798)}
\item{Collection:}{C5, consist of genes annotated by GO Biological Process terms.}
}
}
\keyword{datasets}
