% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MSigDB_v6.0_C5_mouse}
\alias{MSigDB_v6.0_C5_mouse}
\title{MSigDB_v6.0_C5_mouse}
\format{Matrix containing 4436 biological pathways in rows and 18360 genes in columns.}
\source{
\url{http://software.broadinstitute.org/gsea/msigdb}
}
\usage{
data(MSigDB_v6.0_C5_mouse)
}
\description{
A matrix containing mouse gene sets from the MSigDB 6.0 data base (\url{http://software.broadinstitute.org/gsea/msigdb}).
\describe{
\item{species:}{Mus musculus}
\item{Gene IDs:}{Gene symbol/name (i.e. FOXA2)}
\item{Collection:}{C5, consist of genes annotated by GO Biological Process terms.}
}
}
\keyword{datasets}
