% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactomeGS}
\alias{reactomeGS}
\title{reactomeGS}
\format{Rows contain biological pathways and columns contain genes (ensemble IDs).}
\usage{
data(reactomeGS)
}
\description{
A matrix containing human gene sets from the Reactome v59 Pathway data base (https://reactome.org).
\describe{
\item{specie:}{Homo sapiens}
\item{Gene IDs:}{Ensembl IDs from version 75 (i.e. ENSG00000125798)}
}
}
\keyword{datasets}
