% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scMT_data}
\alias{scMT_data}
\title{scMT_data}
\format{A MultiAssayExperiment containing four Experiments:
\itemize{
  \item{RNA expression: }{ExpressionSet with normalized expression values of most variable genes, p=5000}
  \item{Met Enhancers: }{Methylation values at enhancers, p=5000}
  \item{Met CpG Islands: }{Methylation values at CpG Islands, p=5000}
  \item{Met Promoters: }{Methylation values at Promoters, p=5000}
}}
\usage{
data(scMT_data)
}
\description{
A MultiAssayExperiment containing data from a single cell multi-omics study on mESCs.
}
\keyword{datasets}
