% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CLL_data}
\alias{CLL_data}
\title{CLL_data}
\format{A list containing omic measurements for n=200 patient samples that are used as input data for MOFA
\itemize{
  \item{mRNA: }{normalized expression values of most variable genes, p=5000}
  \item{Methylation: }{methylation M-values in most varibale CpG sites, p=4248}
  \item{Drugs: }{viability values in response t different drugs adn concentrations, p=310}
  \item{Mutations: }{Mutation status for selected genes, p=69}
}}
\usage{
data(CLL_data)
}
\description{
A list containing different omic measurements for CLL patient samples
}
\keyword{datasets}
