% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactomeGS}
\alias{reactomeGS}
\title{reactomeGS}
\format{A matrix containing feature binary membership indictors for different genes (in columns)
 and Reactome Gene Sets (in rows).}
\usage{
data(reactomeGS)
}
\description{
A matrix containing feature binary membership indictors for the Reactome Gene Sets.
}
\keyword{datasets}
