% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reactomeGS}
\alias{reactomeGS}
\alias{MSigDB_v6.0_C2_mouse}
\title{MSigDB_v6.0_C2_mouse}
\format{Rows contain biological pathways and columns contain genes (ensemble IDs).}
\usage{
data(reactomeGS)

data(reactomeGS)
}
\description{
A matrix containing mouse gene sets from the MSigDB data base (http://software.broadinstitute.org/gsea/msigdb).
\describe{
\item{specie:}{Mus musculus}
\item{MSigDB version:}{6.0, which corresponds to Ensembl version 75}
\item{Collection:}{C2, curated gene sets from online pathway databases, publications in PubMed, and knowledge of domain experts.}
}

A matrix containing human gene sets from the Reactome Pathway data base (https://reactome.org).
\describe{
\item{specie:}{Homo sapiens}
\item{Reactome version:}{59, which corresponds to Ensembl version 75}
}
}
\keyword{datasets}
