% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scMT_data}
\alias{scMT_data}
\title{scMT_data}
\format{A MultiAssayExperiment containing four Experiments:
\itemize{
  \item{RNA expression: }{ExpressionSet with normalized expression values of the 5000 most variable genes}
  \item{Met Enhancers: }{Methylation values for 5000 CpGs overlapping enhancer elements}
  \item{Met CpG Islands: }{Methylation values for 5000 CpGs overlapping CpG Islands}
  \item{Met Promoters: }{Methylation values for 5000 CpGs overlapping promoters}
}}
\usage{
data(scMT_data)
}
\description{
A MultiAssayExperiment containing data from a single cell multi-omics study (scMT-seq) on mouse embryonic stem cells (mESCs). \cr
The data set was originally published in \code{https://www.ncbi.nlm.nih.gov/pubmed/26752769}.\cr
The MOFA analysis for this data set can be found in \code{http://msb.embopress.org/content/14/6/e8124}
}
\keyword{datasets}
